backapp/requirements.txt
astroid==2.2.5
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-environ==0.4.5
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
isort==4.3.16
lazy-object-proxy==1.3.1
marshmallow==2.19.2
mccabe==0.6.1
Pillow==5.4.1
# pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.8.1
pycodestyle==2.5.0
pylint==2.3.1
python-dotenv==0.10.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
whitenoise==4.1.2
wrapt==1.11.1
